rag_streamlit/requirements.txt
streamlit==1.35.0
langchain==0.2.16
langchain-community==0.2.16
langchain-groq==0.1.9
chromadb==0.5.5
pypdf==4.2.0
python-dotenv==1.0.1
sentence-transformers==2.7.0

